backend/inspirenet-api-optimized/requirements.txt
# Official InSPyReNet implementation
transparent-background==1.3.4

# Core ML dependencies
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# Image processing dependencies (compatible versions for transparent-background)
opencv-python-headless>=4.9.0.80
numpy>=1.21.0,<1.25.0
Pillow>=9.0.0,<11.0.0
scipy>=1.8.0,<1.12.0

# Effects processing dependencies (Phase 1)
scikit-image>=0.21.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
opencv-contrib-python>=4.9.0
numba>=0.58.0

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
slowapi>=0.1.9  # Rate limiting for API abuse prevention

# Image processing and utilities
albumentations>=1.3.0
imageio>=2.31.0

# Google Cloud dependencies
google-cloud-storage>=2.10.0
google-cloud-logging>=3.8.0

# Performance and monitoring
psutil>=5.9.0
prometheus-client>=0.17.0
httpx>=0.25.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
requests>=2.31.0 